* 'develop' of https://github.com/kminsic/Igo:
  test commit
  add 머지 전 수정
  fix 삭제
  add Post 컬럼
  test git ignore test
  fix 신고기능 수정
  add 좋아요 기능api 1개로, 신고하기 기능 구현
  fix Heart관련 전체 수정
  Delete aws.yml
  fix 수정
  fix 주현 수정
  Update application.properties
  fix 유효성 검사 수정 postservice
  fix 상세 게시물 조회 수정, 소셜로그인 관련 수정, postid삭제
  fix 오타 수정
  fix 오타 수정
  add 좋아요 기능 구현
  fix

# Conflicts:
#	.gitignore
#	src/main/java/com/wak/igo/controller/PostController.java
#	src/main/java/com/wak/igo/domain/Member.java
#	src/main/java/com/wak/igo/domain/MyPost.java
#	src/main/java/com/wak/igo/domain/Post.java
#	src/main/java/com/wak/igo/domain/UserDetailsImpl.java
#	src/main/java/com/wak/igo/dto/request/PostRequestDto.java
#	src/main/java/com/wak/igo/dto/response/PostResponseDto.java
#	src/main/java/com/wak/igo/repository/HeartRepository.java
#	src/main/java/com/wak/igo/repository/MemberRepository.java
#	src/main/java/com/wak/igo/service/HeartService.java
#	src/main/java/com/wak/igo/service/KakaoUserService.java
#	src/main/java/com/wak/igo/service/MyPageService.java
#	src/main/java/com/wak/igo/service/NaverUserService.java
#	src/main/java/com/wak/igo/service/PostService.java
#	src/main/java/com/wak/igo/service/UserDetailsServiceImpl.java
#	src/main/resources/application.properties
#	src/main/resources/aws.yml

+++ src/main/java/com/wak/igo/service/PostService.java
@@ -1,16 +1,13 @@
 package com.wak.igo.service;
 
-import com.wak.igo.domain.Comment;
 import com.wak.igo.domain.Member;
 import com.wak.igo.domain.Post;
 import com.wak.igo.domain.UserDetailsImpl;
 import com.wak.igo.dto.request.InterestedTagDto;
 import com.wak.igo.dto.request.PostRequestDto;
-import com.wak.igo.dto.response.CommentResponseDto;
 import com.wak.igo.dto.response.PostResponseDto;
 import com.wak.igo.dto.response.ResponseDto;
 import com.wak.igo.jwt.TokenProvider;
-import com.wak.igo.repository.CommentRepository;
 import com.wak.igo.repository.MemberRepository;
 import com.wak.igo.repository.PostRepository;
 import io.jsonwebtoken.io.IOException;
@@ -19,8 +16,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
 import java.util.List;
+
 import java.util.Optional;
 
 @Service
@@ -30,7 +27,9 @@ public class PostService {
     private final PostRepository postRepository;
     private final MemberRepository memberRepository;
     private final TokenProvider tokenProvider;
-    private final CommentRepository commentRepository;
+
+
+
 
     //전체 게시글 조회
     @Transactional
@@ -40,7 +39,6 @@ public class PostService {
 
     //게시글 조회 (조회수, 좋아요 , 최신순)
     @Transactional
-
     public ResponseDto<?> getAllGroupPosts(String type) {
         if (type.equals("create")) {
             return ResponseDto.success(postRepository.findAllByOrderByCreatedAtDesc());
@@ -53,7 +51,9 @@ public class PostService {
     }
 
 
-   //상세 페이지 조회
+
+
+    //상세 페이지 조회
     @Transactional
     public ResponseDto<?> getDetail(Long id) {
 
@@ -61,21 +61,8 @@ public class PostService {
         if(null == post) {
             return ResponseDto.fail("NOT_FOUND", "게시글이 존재하지 않습니다.");
         }
-        List<Comment> commentList = commentRepository.findAllByPost(post);
-        List<CommentResponseDto> commentResponseDtoList = new ArrayList<>();
-        for (Comment comment : commentList) {
-            commentResponseDtoList.add(
-                    CommentResponseDto.builder()
-                            .id(comment.getId())
-                            .nickname(comment.getMember().getNickname())
-                            .content(comment.getContent())
-                            .createdAt(comment.getCreatedAt())
-                            .modifiedAt(comment.getModifiedAt())
-                            .build()
-            );
-        }
 
-//        post.add_viewCount();
+        post.add_viewCount();
         return ResponseDto.success(
                 PostResponseDto.builder()
                         .title(post.getTitle())
@@ -83,15 +70,14 @@ public class PostService {
                         .viewCount(post.getViewCount())
                         .heartNum(post.getHeartNum())
 //                        .mapData(post.getMapData())
-                        //신고하기 기능 구현 x
-//                        .report(0)
-//                        .tag(post.getTag())
-                        .nickname(post.getMember().getNickname())
+                        .reportNum(0)
+                        .tags(post.getTags())
                         .profile(post.getMember().getProfileImage())
-                        .commentResponseDtoList(commentResponseDtoList)
+                        .nickname(post.getMember().getNickname())
                         .createdAt(post.getCreatedAt())
                         .modifiedAt(post.getModifiedAt())
                         .build());
+
     }
 
 
@@ -108,44 +94,39 @@ public class PostService {
         memberRepository.save(member);
         return ResponseDto.success("저장 완료");
     }
-
-
     //게시글 생성
     @Transactional
 
     public ResponseDto<?> createPost(PostRequestDto postRequestDto, HttpServletRequest request) throws IOException {
 
         Member member = validateMember(request);
+
         if (null == member){
             return ResponseDto.fail("INVALID TOKEN", "TOKEN이 유효하지않습니다");
         }
 
-//        List<MapData> addressList = new ArrayList<>();
         Post post = Post.builder()
                 .member(member)
                 .title(postRequestDto.getTitle())
                 .content(postRequestDto.getContent())
 //                .mapData(postRequestDto.getMapData())
-//                .tag(postRequestDto.getTag())
+                .tags(postRequestDto.getTags())
                 .heartNum(0)
                 .viewCount(0)
-                //신고하기 기능 구현 x
-//                .report(0)
+                .reportNum(0)
                 .build();
         postRepository.save(post);
 
         return ResponseDto.success(
-                //원하시면 추가
                 PostResponseDto.builder()
 
                         .title(postRequestDto.getTitle())
                         .content(postRequestDto.getContent())
 //                        .mapData(postRequestDto.getMapData())
-//                        .tag(postRequestDto.getTag())
+                        .tags(postRequestDto.getTags())
+                        .reportNum(0)
                         .viewCount(0)
                         .heartNum(0)
-                        //신고하기 기능 구현 x
-//                .report(0)
                         .createdAt(post.getCreatedAt())
                         .modifiedAt(post.getModifiedAt())
                         .build());
@@ -154,12 +135,22 @@ public class PostService {
     //    게시글 수정
     @Transactional
     public ResponseDto<?> updatePost(
-            Long id, PostRequestDto requestDto) throws IOException {
-        Post post = isPresentPost(id);
+            Long id, PostRequestDto requestDto,HttpServletRequest request) throws IOException {
 
+        ResponseDto<?> chkResponse = validateCheck(request);
+        if (!chkResponse.isSuccess())
+            return chkResponse;
+
+        Member member = (Member) chkResponse.getData();
+        // 유저 테이블에서 유저객체 가져오기
+        Member updateMember = memberRepository.findByNickname(member.getNickname()).get();
+
+        Post post = isPresentPost(id);
         if (null == post) {
             return ResponseDto.fail("NOT_FOUND", "게시글이 존재하지 않습니다.");
         }
+        if (post.validateMember(updateMember))
+            return ResponseDto.fail("작성자가 아닙니다.","작성자가 아닙니다.");
 
         post.update(requestDto);
         return ResponseDto.success("update success");
@@ -168,14 +159,26 @@ public class PostService {
 
     //게시글 삭제
     @Transactional
-    public ResponseDto<?> deletePost(Long id) {
+    public ResponseDto<?> deletePost(Long id, HttpServletRequest request) {
+
+        ResponseDto<?> chkResponse = validateCheck(request);
+        if (!chkResponse.isSuccess())
+            return chkResponse;
+
+        Member member = (Member) chkResponse.getData();
+        Member updateMember = memberRepository.findByNickname(member.getNickname()).get();
+
         Post post = isPresentPost(id);
+        if (post == null)
+            return ResponseDto.fail("글 삭제에 실패하였습니다. (NOT_EXIST)", "글 삭제에 실패하였습니다. (NOT_EXIST)");
+
+        if (post.validateMember(updateMember))
+            return ResponseDto.fail("작성자가 아닙니다.", "작성자가 아닙니다.");
         postRepository.delete(post);
         return ResponseDto.success("Success");
 
     }
 
-    //
     @Transactional(readOnly = true)
     public Post isPresentPost(Long id) {
         Optional<Post> optionalPost = postRepository.findById(id);
@@ -189,13 +192,20 @@ public class PostService {
         }
         return tokenProvider.getMemberFromAuthentication().getMember();
     }
+    private ResponseDto<?> validateCheck(HttpServletRequest request) {
+        // RefreshToken 및 Authorization 유효성 검사
+        if (null == request.getHeader("RefreshToken")
+                || null == request.getHeader("Authorization")) {
+            return ResponseDto.fail("로그인이 필요합니다.","로그인이 필요합니다.");
+        }
+        Member member = validateMember(request);
+        // token 정보 유효성 검사
+        if (null == member) {
+            return ResponseDto.fail("Token이 유효하지 않습니다.","Token이 유효하지 않습니다.");
+        }
+        return ResponseDto.success(member);
+    }
+
+
 }
 
-//신고기능 미구현
-//    @Transactional
-//    public ResponseDto<?> getReport(Long id) {
-//        if(report >= 50 ){
-//            postRepository.delete(id);
-//        }
-//
-//    }
